pyproject.toml
[tool.poetry]
name = "dupehunter"
version = "0.1.0"
description = "A Python package for detecting and managing duplicate image files."
authors = ["Fredrik Rundgren <fredrik@leipt.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/frecke/dupehunter"
keywords = ["duplicate", "image", "checksum", "deduplication"]

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^9.0"
aiofiles = "^22.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-cov = "^6.0.0"
black = "^23.0"
flake8 = "^6.0"
mypy = "^1.0"
isort = "^5.0"
pre-commit = "^4.0.1"
pytest-mock = "^3.14.0"
pytest-xdist = "^3.6.1"
types-pillow = "^10.2.0.20240822"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
exclude = '''
/(
    \.git
  | \.venv
  | \.mypy_cache
  | \.pytest_cache
  | build
  | dist
)/
'''
